Added a XY pad to the Delay Pluging. also added a volume control, and smothed the delay time input  changes, to improve audio.

+++ plugins/Delay/StereoDelay.cpp
@@ -74,7 +74,7 @@ void StereoDelay::tick( sampleFrame frame )
 	m_buffer[m_index][0] += frame[0] * m_feedback;
 	m_buffer[m_index][1] += frame[1] * m_feedback;
 
-	m_index = ( m_index + 1) % m_maxLength;
+	m_index = ( m_index + 1) % (int) m_maxLength;
 }
